pyproject.toml
[project]
name = "ui-distiller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "globus-sdk>=4.3.0",
    "json-repair>=0.54.3",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "pyarrow>=22.0.0",
    "scipy>=1.16.3",
    "tiktoken>=0.12.0",
]
